feat: improve template
