+++ templates/interchaintestv8/chainconfig/files/chainconfig/chain_config.go.plush
@@ -3,8 +3,6 @@ package chainconfig
 import (
 	interchaintest "github.com/strangelove-ventures/interchaintest/v8"
 	"github.com/strangelove-ventures/interchaintest/v8/ibc"
-
-	"<%= ModulePath %>/e2esuite"
 )
 
 var DefaultChainSpecs = []*interchaintest.ChainSpec{
@@ -26,7 +24,7 @@ var DefaultChainSpecs = []*interchaintest.ChainSpec{
 			Denom:          "stake",
 			GasPrices:      "0.00stake",
 			GasAdjustment:  1.3,
-			EncodingConfig: e2esuite.EncodingConfig(),
+			EncodingConfig: EncodingConfig(),
 			TrustingPeriod: "508h",
 			NoHostMount:    false,
 		},
@@ -49,7 +47,7 @@ var DefaultChainSpecs = []*interchaintest.ChainSpec{
 			Denom:          "stake",
 			GasPrices:      "0.00stake",
 			GasAdjustment:  1.3,
-			EncodingConfig: e2esuite.EncodingConfig(),
+			EncodingConfig: EncodingConfig(),
 			TrustingPeriod: "508h",
 			NoHostMount:    false,
 		},
