+++ templates/interchaintestv8/files/README.md.plush
@@ -1,0 +1,16 @@
+# End to End Testing Suite with Interchaintest
+
+The e2e tests are built using the [interchaintest](https://github.com/strangelove-ventures/interchaintest) library by Strangelove. It runs multiple docker container validators, and lets you test IBC enabled smart contracts.
+
+These end to end tests are designed to run in the ci, but you can also run them locally.
+
+## Running the tests locally
+
+To run the tests locally, run the following commands from this directory:
+
+```text
+cd interchaintest/
+go test -v . -run TestWithBasicTestSuite -testify.m $TEST_NAME
+```
+
+where `$TEST_NAME` is one of the test names of the `BasicTestSuite` in `basic_test.go`.
